+++ tomcat/src/main/java/com/techcourse/handler/StaticResourceRequestHandler.java
@@ -1,12 +1,13 @@
 package com.techcourse.handler;
 
-import com.techcourse.http.ContentType;
-import com.techcourse.http.HttpCookie;
-import com.techcourse.http.HttpVersion;
-import com.techcourse.http.request.HttpRequest;
-import com.techcourse.http.response.HttpResponse;
-import com.techcourse.http.response.ResponseBody;
+import com.techcourse.handler.core.AbstractRequestHandler;
 import com.techcourse.util.FileUtil;
+import org.apache.coyote.http.ContentType;
+import org.apache.coyote.http.HttpCookie;
+import org.apache.coyote.http.HttpVersion;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.response.ResponseBody;
 
 public class StaticResourceRequestHandler extends AbstractRequestHandler {
     private final HttpVersion httpVersion;
