+++ tomcat/src/main/java/com/techcourse/handler/core/RequestHandler.java
@@ -1,0 +1,8 @@
+package com.techcourse.handler.core;
+
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+
+public interface RequestHandler {
+    HttpResponse service(HttpRequest httpRequest) throws Exception;
+}
