+++ tomcat/src/main/java/org/apache/coyote/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package com.techcourse.http;
+package org.apache.coyote.http;
 
 import com.techcourse.exception.NotFoundException;
 import java.util.Arrays;
