+++ tomcat/src/main/java/org/apache/coyote/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package com.techcourse.http;
+package org.apache.coyote.http;
 
 
 import com.techcourse.exception.NotFoundException;
