+++ tomcat/src/main/java/org/apache/coyote/http/request/RequestBody.java
@@ -1,4 +1,4 @@
-package com.techcourse.http.request;
+package org.apache.coyote.http.request;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.util.Map;
