+++ tomcat/src/main/java/org/apache/coyote/http/response/Location.java
@@ -1,4 +1,4 @@
-package com.techcourse.http.response;
+package org.apache.coyote.http.response;
 
 public class Location {
