+++ tomcat/src/test/java/com/techcourse/handler/LoginRequestHandlerTest.java
@@ -6,14 +6,14 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.techcourse.exception.NotFoundException;
 import com.techcourse.exception.UncheckedServletException;
-import com.techcourse.http.HttpVersion;
-import com.techcourse.http.request.HttpRequest;
-import com.techcourse.http.request.RequestBody;
-import com.techcourse.http.request.RequestHeader;
-import com.techcourse.http.response.HttpResponse;
-import com.techcourse.http.session.Session;
-import com.techcourse.http.session.SessionRepository;
 import java.util.List;
+import org.apache.coyote.http.HttpVersion;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.request.RequestBody;
+import org.apache.coyote.http.request.RequestHeader;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.session.Session;
+import org.apache.coyote.http.session.SessionRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
