+++ tomcat/src/test/java/com/techcourse/http/common/ContentTypeTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.techcourse.exception.NotFoundException;
-import com.techcourse.http.ContentType;
+import org.apache.coyote.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
