+++ tomcat/src/test/java/com/techcourse/http/common/HttpCookieTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.techcourse.exception.UncheckedServletException;
-import com.techcourse.http.HttpCookie;
+import org.apache.coyote.http.HttpCookie;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import org.junit.jupiter.api.DisplayName;
