+++ tomcat/src/test/java/com/techcourse/http/common/HttpMethodTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.techcourse.exception.NotFoundException;
-import com.techcourse.http.HttpMethod;
+import org.apache.coyote.http.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
