+++ tomcat/src/test/java/com/techcourse/http/request/RequestHeaderTest.java
@@ -7,6 +7,7 @@ import com.techcourse.exception.UncheckedServletException;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import org.apache.coyote.http.request.RequestHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
